Add files via upload
Mapping of metabolites from Sco4 to iKS1317 in table: COMPARE_METABOLITES_iKS1317_Sco4.xlsx
Metabolites marked in red are involved in both models, Brown are manually checcked (Chemical formula) and they present the same metabolites in both models-although they have slightly different IDs in those two models, blue are in iKS1317 but not in Sco4 and green in bold are in Sco4 but not in iKS1317. So the metabolites, marked in green should be added to iKS1317-those metabolites I also mapped in KEGG IDs. Each metaboites have A-S columns